Implement catalyst_vars parameter
This allows the user to restrict which variables are passed to Catalyst.
In theory this could be determined from the Catalyst script but the
functions to determine this don't seem to work well with newer Catalyst
scripts (i.e. those generated by v5.9 or later) and always request all
variables.