Initialize `vtkUniformGrid`s from `vtkAMRBox`s
This shifts the set `origin` of the `vtkUniformGrid`s from the
`global_origin` at (0,0,0) (in Chombo coordinates) to that of the low
corner of the box. This seems to disagree with what is done in the
ParaView Catalyst CxxOverlappingAMRExample but makes the outputted VTK
XML files appear more sensible when loaded in ParaView.