tread more carefully with automatic LUT ranges
In most cases, a recorded catalyst script completely specifies
the LUT range within the python script and all ranks use this.
In my case I was altering the default one that Catalyst/ParaView
builds when not specified and ran into a bug where each rank
creates its own LUT, the ranges of which can vary. With these
scripts I more carefully set up the LUT and thus no longer
need to hardcode the ranges to the data in the c++ level to work
around the obscure bug. In one case we simply completely specify
the LUT, in the other we fixup the default one.